RestaurantsTabs implemented
