+++ src/components/restaurants/component.jsx
@@ -1,26 +1,13 @@
-import { useSelector } from 'react-redux';
-import { Restaurant } from '../restaurant/component';
-import { Tab } from '../tab/component';
 import { useState } from 'react';
-import { RestaurantTab } from '../tab/container';
+import { RestaurantsTabs } from '../restaurants-tabs/components';
+import { Restaurant } from '../restaurant/component';
 
 export const Restaurants = () => {
-    const restaurantsIds = useSelector(
-        (state) => state.restaurant.ids
-    );
-
-    const [activeRestaurantId, setActiveRestaurantId] = useState(restaurantsIds[0]);
+    const [activeRestaurantId, setActiveRestaurantId] = useState();
 
     return (
         <div>
-            <div>
-                {restaurantsIds.map((id) => (
-                    <RestaurantTab
-                        restaurantId={id}
-                        setActiveTab={() => setActiveRestaurantId(id)}
-                        active={id === activeRestaurantId} />
-                ))}
-            </div>
+            <RestaurantsTabs activeRestaurantId={activeRestaurantId} setActiveRestaurantId={setActiveRestaurantId} />
             <Restaurant restaurantId={activeRestaurantId} />
         </div>
     );
